Data Sync done
-Data is automatically kept in sync between devices via polling parse
every 10 seconds.  This is done using a remote object that is set by
the app whenever a new object is created, modified or deleted and each
“poll” increments a counter before finally deleting the remote object.
If the object is found, a data refresh is performed.  Using a counter
should give any devices currently running the app enough time to
refresh successfully.