Redid data sync
-Redid how the data synchronization worked as it was flaky before and
oftentimes, more than one device would get into a race condition to
update remote data, which could result in the other device that
actually needed to update, not updating.  Each update “object” stored
on parse now uses a unique key that is referenced within the device and
is unique to that device.