Fixed no user bug/flaw in data sync
-Fixed a giant bug in the “add event” activity where regaining network
connection after having been in “offline” mode would cause a crash when
the save button was tapped.  This is because the app is attempting to
set the ACL value for the ParseObject when there wasn’t a currentUser.

-Implemented a counter variable into the data synchronization logic so
that all currently running devices have enough time to grab the new
data before the remote ‘token’ is deleted.  Before, whatever device
grabbed the token first would be the one that would update and then
immediately delete the remote token.  For 2 devices, this would work
fine, but for more than that, the additional devices would not auto
update as there wouldn’t be the remote token to signalize an update is
needed.