'Offline' mode complete
-Completed adding in an ‘offline’ mode which basically just lets the
user into the app without a network connection.  Unfortunately, due to
lack of time, the user can’t really create or manipulate any data while
in this mode.  Added in alerts to notify the user when the network is
restored, and give them the option to log in as well.